[CHORE] #12 - 한 곳에서 데이터 관리하도록 변경
